requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
langchain_core
streamlit